Hangman
This famous game is one of those games that is fun to play digitally. Now and with this simple code, you can play hangman! 
You can also change the hangman art file and be more creative or update the hangman words file to use your own source of vocabularies.
This program is written using while loop, for loops and nested ifs.