caesar_cipher
This is a program to encode/decode the input text ( just the alphabet ) based on the famous Caesar Cipher code. It first asks you to choose an option between decoding or encoding your message, then will ask for your message, and will finally ask for the number of shifts you would like to have for encoding/decoding.