higher_lower_game
This game is a variation of  higher "www.higherlowergame.com" and it uses the same rules and methodologies. But the data are gathered from celebrities' Instagram. So, the goal is for the user to guess the right number of followers in each comparison and you can continue the game as long as you are right.