Pomodoro Timer
This is the famous Pomodoro technique that will help you plan your time and keep up the good work. So, the basic rule of it is that you need to work for 25 minutes on your task followed by a short break of 5. Then, repeat the process 4 times in total which will earn you a long break of 20 minutes. This program will calculate everything automatically for you!