Flashcard
This is a flashcard program for learning new French vocabularies from a list of 100 most common French words.
When shown a new card, you have 4 seconds before the program shows you the back of the card, that is the translation. If you click on the checkmark, you will see another random card pops up and that word will not be shown again. And the same thing happens if you click on the cross mark except you will see the unknown word again until you master it.