Birthday Wisher
This is a simple program that will automate the process of creating a birthday message and send it via email to your loved ones according to the data that you have added in the birthdays.csv file. In order to automatically run the program everyday, I recommend using a website such as www.pythonanywhere.com to run the program everyday.