Amazon Price Tracker
This program will email you the product title you are interested in - based on the link you provide - as well as the price and the link for that product. However, you need to change the block of code to search for the price and the title of the item accordingly.