Cookie Clicker
This is a program that will help you achieve outstanding cookie per second score in "http://orteil.dashnet.org/experiments/cookie/" which is the classic version of the cookie clicker. This program uses selenium module to find and track objects on screen and essentially click on the cookie for the defined duration.